chore: moved trigger segmentation render
