+++ packages/tools/src/eventListeners/segmentation/segmentationRepresentationRemovedEventListener.ts
@@ -1,4 +1,4 @@
-import triggerSegmentationRender from '../../utilities/triggerSegmentationRender';
+import triggerSegmentationRender from '../../utilities/segmentation/triggerSegmentationRender';
 import { SegmentationRepresentationRemovedEventType } from '../../types/EventTypes';
 
 /** A function that listens to the `segmentationRepresentationRemoved` event and triggers
