+++ packages/tools/src/utilities/segmentation/index.ts
@@ -4,6 +4,7 @@ import createMergedLabelmapForIndex from './createMergedLabelmapForIndex';
 import isValidRepresentationConfig from './isValidRepresentationConfig';
 import getDefaultRepresentationConfig from './getDefaultRepresentationConfig';
 import createLabelmapVolumeForViewport from './createLabelmapVolumeForViewport';
+import { triggerSegmentationRender } from './triggerSegmentationRender';
 import {
   getBrushThresholdForToolGroup,
   setBrushThresholdForToolGroup,
@@ -24,4 +25,5 @@ export {
   setBrushSizeForToolGroup,
   getBrushThresholdForToolGroup,
   setBrushThresholdForToolGroup,
+  triggerSegmentationRender,
 };
